requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Pillow==10.1.0
qrcode==7.4.2
Werkzeug==3.0.1
flask-cors

